+++ src/main/java/com/anon/backend/entity/Message.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 * 
 *
 * @author anon
-* @since 2024-10-30
+* @since 2025-01-14
 */
 @Getter
 @Setter
@@ -23,11 +23,11 @@ public class Message implements Serializable {
   private static final long serialVersionUID = 1L;
 
   @TableId(value = "id", type = IdType.AUTO)
-  private Integer id;
+  private Long id;
 
-  private Integer sender;
+  private Long sender;
 
-  private Integer receiver;
+  private Long receiver;
 
   private String content;
